removed absolute position from header and fixed icons footer
