+++ src/components/Header/Header.style.js
@@ -1,0 +1,33 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+    headerContainer:{
+        height: 80,
+        flexDirection: 'row',
+        justifyContent: "flex-start",
+        flex: 1,
+    }, 
+    headerLogo:{
+        marginTop: 40,
+    },
+    headerTxt:{
+        color: 'white',
+        fontSize: 20,
+        fontWeight: '700',
+        marginLeft: 15,
+    },
+    notIconContainer:{
+        marginTop: 40,
+        marginLeft: 170,
+    },
+    profileIconContainer:{
+        marginTop: 40,
+        marginLeft: 25
+    },
+    notificationIcon:{
+        
+    },
+    profileIcon:{
+
+    }
+});
