+++ src/components/Homepage.js
@@ -6,18 +6,19 @@ import {
     Button
 } from 'react-native';
 
-import Header from './Header';
-import Footer from './Footer';
+import Header from './Header/Header';
+import Footer from './Footer/Footer';
 
 class Homepage extends Component{
     
     render(){
         return(
             <View style={styles.homepageContainer}>
+            <Header/>
                 <ScrollView style ={styles.bodyContainer}> 
                   
                 </ScrollView>
-                <Header/>
+                
                 <Footer/>
             </View>
