+++ src/components/ImageSelector/ImageSelector.style.js
@@ -16,6 +16,7 @@ export default StyleSheet.create({
         width: 150, 
         height: 150,
         borderRadius: 75,
-    }
+    }, 
+   
 
 });
